refactor: renomeia nomes de componentes já criados para o padrão dash-case.
